requirements.txt
GitPython==3.1.27
Pillow==9.2.0
accelerate==0.14.0
albumentations==1.3.0
basicsr==1.4.2
bitsandbytes==0.35.0
clean-fid==0.1.29
diffusers==0.7.2
einops==0.4.1
fairscale==0.4.9
fastapi==0.87.0
font-roboto
fonts
ftfy==6.1.1
gfpgan==1.3.8
gradio==3.9
inflection==0.5.1
jsonmerge==1.8.0
kornia==0.6.7
lark==1.1.2
modelcards==0.1.6
numpy==1.23.3
omegaconf==2.2.3
piexif==1.1.3
pytorch_lightning==1.7.6
realesrgan==0.3.0
resize-right==0.0.2
scikit-image==0.19.2
tensorboard==2.11.0
timm==0.6.7
torch
torch-ema
ninja
trimesh
PyMCubes
rich
xatlas
imageio
imageio-ffmpeg
accelerate
torchdiffeq==0.2.3
tqdm==4.64.1
transformers==4.21.0
git+https://github.com/toanngosy/stable-dreamfusion@setup.py
git+https://github.com/NVlabs/nvdiffrast
